+++ Source/AccelByteUe4Sdk/Private/Api/AccelByteGameTelemetryApi.cpp
@@ -9,12 +9,14 @@
 #include "Core/AccelByteHttpRetryScheduler.h"
 #include "JsonUtilities.h"
 #include "Core/AccelByteSettings.h"
+#include <AccelByteUe4SdkModule.h>
+#include <Core/AccelByteUtilities.h>
 
 namespace AccelByte
 {
 namespace Api
 {
-GameTelemetry::GameTelemetry(Credentials const& InCredentialsRef
+GameTelemetry::GameTelemetry(Credentials& InCredentialsRef
 	, Settings const& InSettingsRef
 	, FHttpRetryScheduler& InHttpRef)
 	: HttpRef{InHttpRef}
@@ -22,11 +24,17 @@ GameTelemetry::GameTelemetry(Credentials const& InCredentialsRef
 	, SettingsRef{InSettingsRef}
 	, ShuttingDown(false)
 {
+	CredentialsRef.OnLoginSuccess().AddRaw(this, &GameTelemetry::OnLoginSuccess);
 }
 
 GameTelemetry::~GameTelemetry()
 {
 	Shutdown();
+}
+
+void GameTelemetry::OnLoginSuccess(FOauth2Token const& Response)
+{
+	LoadCachedEvents();
 }
 
 void GameTelemetry::SetBatchFrequency(FTimespan Interval)
@@ -63,11 +71,13 @@ void GameTelemetry::Send(FAccelByteModelsTelemetryBody TelemetryBody, FVoidHandl
 
 	if (ImmediateEvents.Contains(TelemetryBody.EventName))
 	{
-		SendProtectedEvents({ TelemetryBody }, OnSuccess, OnError);
+		SendProtectedEvents({ MakeShared<FAccelByteModelsTelemetryBody>(TelemetryBody) }, OnSuccess, OnError);
 	}
 	else
 	{
-		JobQueue.Enqueue(TTuple<FAccelByteModelsTelemetryBody, FVoidHandler, FErrorHandler>{ TelemetryBody, OnSuccess, OnError });
+		TSharedPtr<FAccelByteModelsTelemetryBody> TelemetryPtr = MakeShared<FAccelByteModelsTelemetryBody>(TelemetryBody);
+		JobQueue.Enqueue(TTuple<TSharedPtr<FAccelByteModelsTelemetryBody>, FVoidHandler, FErrorHandler>{ TelemetryPtr, OnSuccess, OnError });
+		AppendEventToCache(TelemetryPtr);
 		if (bTelemetryJobStarted == false)
 		{
 			bTelemetryJobStarted = true;
@@ -108,12 +118,12 @@ bool GameTelemetry::PeriodicTelemetry(float DeltaTime)
 	{
 		FReport::Log(FString(__FUNCTION__));
 
-		TArray<FAccelByteModelsTelemetryBody> TelemetryBodies;
+		TArray<TSharedPtr<FAccelByteModelsTelemetryBody>> TelemetryBodies;
 		TArray<FVoidHandler> OnSuccessCallbacks;
 		TArray<FErrorHandler> OnErrorCallbacks;
 		while (!JobQueue.IsEmpty())
 		{
-			TTuple<FAccelByteModelsTelemetryBody, FVoidHandler, FErrorHandler> DequeueResult;
+			TTuple<TSharedPtr<FAccelByteModelsTelemetryBody>, FVoidHandler, FErrorHandler> DequeueResult;
 			if (JobQueue.Dequeue(DequeueResult))
 			{
 				TelemetryBodies.Add(DequeueResult.Get<0>());
@@ -121,12 +131,14 @@ bool GameTelemetry::PeriodicTelemetry(float DeltaTime)
 				OnErrorCallbacks.Add(DequeueResult.Get<2>());
 			}
 		}
+		EventPtrArray.Empty();
 
 		SendProtectedEvents(
 			TelemetryBodies,
 			FVoidHandler::CreateLambda(
-				[OnSuccessCallbacks]()
+				[this, OnSuccessCallbacks]()
 				{
+					RemoveEventsFromCache();
 					for (auto& OnSuccessCallback : OnSuccessCallbacks)
 					{
 						OnSuccessCallback.ExecuteIfBound();
@@ -144,7 +156,7 @@ bool GameTelemetry::PeriodicTelemetry(float DeltaTime)
 	return true;
 }
 
-void GameTelemetry::SendProtectedEvents(TArray<FAccelByteModelsTelemetryBody> Events, FVoidHandler const& OnSuccess, FErrorHandler const& OnError)
+void GameTelemetry::SendProtectedEvents(TArray<TSharedPtr<FAccelByteModelsTelemetryBody>> const& Events, FVoidHandler const& OnSuccess, FErrorHandler const& OnError)
 {
 	if (ShuttingDown)
 	{
@@ -164,10 +176,10 @@ void GameTelemetry::SendProtectedEvents(TArray<FAccelByteModelsTelemetryBody> Ev
 	for (auto const& Event : Events)
 	{
 		TSharedPtr<FJsonObject> JsonObject = MakeShared<FJsonObject>();
-		JsonObject->SetStringField("EventNamespace", Event.EventNamespace);
-		JsonObject->SetStringField("EventName", Event.EventName);
-		JsonObject->SetObjectField("Payload", Event.Payload);
-		JsonObject->SetStringField("EventTimestamp", Event.EventTimestamp.ToIso8601());
+		JsonObject->SetStringField("EventNamespace", Event->EventNamespace);
+		JsonObject->SetStringField("EventName", Event->EventName);
+		JsonObject->SetObjectField("Payload", Event->Payload);
+		JsonObject->SetStringField("EventTimestamp", Event->EventTimestamp.ToIso8601());
 
 		JsonArray.Add(MakeShared<FJsonValueObject>(JsonObject));
 	}
@@ -183,6 +195,128 @@ void GameTelemetry::SendProtectedEvents(TArray<FAccelByteModelsTelemetryBody> Ev
 	Request->SetContentAsString(Content);
 
 	HttpRef.ProcessRequest(Request, CreateHttpResultHandler(OnSuccess, OnError), FPlatformTime::Seconds());
+}
+
+void GameTelemetry::LoadCachedEvents()
+{
+	FString TelemetryKey = GetTelemetryKey();
+	if (TelemetryKey.IsEmpty())
+	{
+		return;
+	}
+	IAccelByteUe4SdkModuleInterface::Get().GetLocalDataStorage()->GetItem(TelemetryKey, THandler<TPair<FString, FString>>::CreateLambda(
+		[this](TPair<FString, FString> Pair)
+		{
+			if (Pair.Key.IsEmpty() || Pair.Value.IsEmpty())
+			{
+				return;
+			}
+			TArray<TSharedPtr<FAccelByteModelsTelemetryBody>> EventList;
+			if (EventsJsonToArray(Pair.Value, EventList))
+			{
+				SendProtectedEvents(
+					EventList,
+					FVoidHandler::CreateLambda([this]()
+					{
+						RemoveEventsFromCache();
+					}),
+					FErrorHandler::CreateLambda([](int32 ErrorCode, const FString& ErrorMessage) {}));
+			}
+		}));
+}
+
+void GameTelemetry::AppendEventToCache(TSharedPtr<FAccelByteModelsTelemetryBody> Telemetry)
+{
+	FString TelemetryKey = GetTelemetryKey();
+	if (TelemetryKey.IsEmpty())
+	{
+		return;
+	}
+	FString TelemetryValues;
+	EventPtrArray.Add(Telemetry);
+	JobArrayQueueAsJsonString(TelemetryValues);
+	IAccelByteUe4SdkModuleInterface::Get().GetLocalDataStorage()->SaveItem(
+		TelemetryKey,
+		TelemetryValues,
+		THandler<bool>::CreateLambda([](bool IsSuccess){}));
+}
+
+void GameTelemetry::RemoveEventsFromCache()
+{
+	FString TelemetryKey = GetTelemetryKey();
+	if (TelemetryKey.IsEmpty())
+	{
+		return;
+	}
+	IAccelByteUe4SdkModuleInterface::Get().GetLocalDataStorage()->DeleteItem(
+		TelemetryKey,
+		FVoidHandler::CreateLambda([](){}));
+}
+
+bool GameTelemetry::JobArrayQueueAsJsonString(FString& OutJsonString)
+{
+	TSharedRef<FJsonObject> TelemetryObj = MakeShared<FJsonObject>();
+	TArray<TSharedPtr<FJsonValue>> EventsObjArray;
+	for (auto& EventPtr : EventPtrArray)
+	{
+		TSharedPtr<FJsonObject> JsonObj = MakeShared<FJsonObject>();
+		JsonObj->SetStringField("EventName", EventPtr->EventName);
+		JsonObj->SetStringField("EventNamespace", EventPtr->EventNamespace);
+		JsonObj->SetObjectField("Payload", EventPtr->Payload);
+		JsonObj->SetNumberField("EventTimestamp", EventPtr->EventTimestamp.ToUnixTimestamp());
+		TSharedRef<FJsonValueObject> JsonValue = MakeShared<FJsonValueObject>(JsonObj);
+		EventsObjArray.Add(JsonValue);
+	}
+	TelemetryObj->SetArrayField("telemetry", EventsObjArray);
+	TSharedRef<TJsonWriter<>> Writer = TJsonWriterFactory<>::Create(&OutJsonString);
+	if (!FJsonSerializer::Serialize(TelemetryObj, Writer))
+	{
+		return false;
+	}
+	return true;
+}
+
+bool GameTelemetry::EventsJsonToArray(FString& InJsonString, TArray<TSharedPtr<FAccelByteModelsTelemetryBody>>& OutArray)
+{
+	TSharedPtr<FJsonObject> JsonObjectPtr{ MakeShared<FJsonObject>() };
+
+	if (!FJsonSerializer::Deserialize(TJsonReaderFactory<>::Create(InJsonString), JsonObjectPtr))
+	{
+		return false;
+	}
+
+	TArray<TSharedPtr<FJsonValue>> const* ArrayValue;
+	if (!JsonObjectPtr->TryGetArrayField("telemetry", ArrayValue))
+	{
+		return false;
+	}
+
+	for (auto& ArrayItem : *ArrayValue)
+	{
+		auto& JsonObj = ArrayItem->AsObject();
+		FString EventName, EventNamespace;
+		TSharedPtr<FJsonObject> const* Payload = nullptr;
+		int32 EventTimeStamp = 0;
+		FAccelByteModelsTelemetryBody TelemetryBody;
+		TelemetryBody.EventName = JsonObj->GetStringField("EventName");
+		TelemetryBody.EventNamespace = JsonObj->GetStringField("EventNamespace");
+		TelemetryBody.Payload = JsonObj->GetObjectField("Payload");
+		EventTimeStamp = JsonObj->GetIntegerField("EventTimestamp");
+		TelemetryBody.EventTimestamp = FDateTime::FromUnixTimestamp(EventTimeStamp);
+		OutArray.Add(MakeShared<FAccelByteModelsTelemetryBody>(TelemetryBody));
+	}
+	return true;
+}
+
+FString GameTelemetry::GetTelemetryKey()
+{
+	FString UserId = *CredentialsRef.GetUserId();
+	if (UserId.IsEmpty())
+	{
+		return UserId;
+	}
+	FString TelemetryKey = FString::Printf(TEXT("TELEMETRY_%s"), *UserId);
+	return TelemetryKey;
 }
 
 }